Changed id-type in customer

+++ src/main/java/com/knowit/gb/example/demo/controller/CustomerController.java
@@ -1,5 +1,6 @@
 package com.knowit.gb.example.demo.controller;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.http.ResponseEntity;
@@ -47,7 +48,7 @@ public class CustomerController {
 	}
 
 	@GetMapping("/zipcode/{zipcode}")
-	public ResponseEntity<List<Customer>> getCustomerByZipCodeContains(@PathVariable int zipCode) {
+	public ResponseEntity<List<Customer>> getCustomerByZipCodeContains(@PathVariable BigInteger zipCode) {
 		return ResponseEntity.ok(customerRepository.findByZipCodeContains(zipCode));
 	}
